1. Add Pause and Resume test cases 2. Add one failpoint of commit error during admin pause/resume/cancel
